+++ examples/ex_preload_replay_buffer.bash
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
 
-
 ## Random seed to use for both the environment and agent (-1 for random)
 SEED="123"
 
@@ -19,6 +18,7 @@ ENV_ID="Grasp-OctreeWithColor-Gazebo-v0"
 ROBOT_MODEL="panda"
 # ROBOT_MODEL="ur5_rg2"
 # ROBOT_MODEL="kinova_j2s7s300"
+# ROBOT_MODEL="lunalab_summit_xl_gen"
 
 ## Algorithm to use (might not matter too much as long as it is off-policy)
 # ALGO="sac"
@@ -38,12 +38,10 @@ EXTRA_ARGS=""
 ########################################################################################################################
 
 ## Spawn ign_moveit2 subprocess in background, while making sure to forward termination signals
-IGN_MOVEIT2_CMD="ros2 launch drl_grasping ign_moveit2_headless.launch.py"
-if [ "$ROBOT_MODEL" = "ur5_rg2" ]; then
-    IGN_MOVEIT2_CMD="ros2 launch drl_grasping ign_moveit2_headless_ur5_rg2.launch.py"
-fi
+IGN_MOVEIT2_CMD="ros2 launch drl_grasping sim.launch.py robot_model:=${ROBOT_MODEL} enable_rviz:=false"
 if [ "$ROBOT_MODEL" = "kinova_j2s7s300" ]; then
-    IGN_MOVEIT2_CMD="ros2 launch drl_grasping ign_moveit2_headless_kinova_j2s7s300.launch.py"
+    # Robot name for `kinova_j2s7s300` is different
+    IGN_MOVEIT2_CMD="${IGN_MOVEIT2_CMD} robot_name:=j2s7s300"
 fi
 echo "Launching ign_moveit2 in background:"
 echo "${IGN_MOVEIT2_CMD}"